Adds cartID to bills and attach bill items to the Lof struct
the shopping cart ID is intended to ensure that invoices cannot be accidentally billed more than once. Adding the goods as a QVariant to the log struct is so that other plugins have access to this information to process it externally. Furthermore, the ObjectID of the shopping cart is stored in the log to have a mapping between the shopping cart and the log entry in the future.